// extension/src/bitstream/commands/command-types.ts
export type BitstreamCommandName =
  | "HELLO_REQ"
  | "PING_REQ"
  | "CAPS_REQ"
  | "STATUS_REQ"
  | "STREAM_PAUSE_REQ"
  | "STREAM_RESUME_REQ"
  | "SENSOR_CFG_GET_REQ"
  | "SENSOR_CFG_SET_REQ"
  | "SENSOR_REINIT_REQ"
  | "BMI270_MODE_SET_REQ"
  | "BMI270_MODE_GET_REQ"
  | "BMI270_FUSION_FEED_SET_REQ"
  | "BMI270_FUSION_FEED_GET_REQ"
  | "LOG_LEVEL_GET_REQ"
  | "LOG_LEVEL_SET_REQ"
  | "DIAG_GET_SNAPSHOT_REQ"
  | "DIAG_STREAM_START_REQ"
  | "DIAG_STREAM_STOP_REQ"
  | "DIAG_GET_TASK_TABLE_REQ"
  | "DIAG_SET_TASK_PRIORITY_REQ"
  | "DIAG_TASK_STREAM_CONFIG_SET_REQ"
  | "DIAG_TASK_STREAM_FILTER_BEGIN_REQ"
  | "DIAG_TASK_STREAM_FILTER_CHUNK_REQ"
  | "DIAG_TASK_STREAM_FILTER_COMMIT_REQ"
  | "DIAG_TASK_STREAM_RESYNC_NOW_REQ"
  | "WIFI_CONNECT_REQ"
  | "WIFI_DISCONNECT_REQ"
  | "WIFI_STATUS_POLL_REQ"
  | "WIFI_SCAN_ALL_REQ"
  | "WIFI_SCAN_SSID_REQ";

export interface BitstreamCommandDefinition {
  name: BitstreamCommandName;
  channel: number;
  commandId: number;
}